+++ yaknawa/server/server_ks.js
@@ -1,0 +1,31 @@
+const express = require("express");
+const bodyParser = require("body-parser");
+const cors = require("cors");
+
+const app = express();
+
+var corsOptions = {
+  origin: "http://localhost:8081",
+};
+
+// Express 앱을 만들기 위해 body-parser와 cors를 app.use를 사용해서 추가함
+app.use(cors(corsOptions));
+app.use(bodyParser.json());
+
+// parse requests of content-type - application/x-www-form-urlencoded
+app.use(bodyParser.urlencoded({ extended: true }));
+
+// simple route
+app.get("/", (req, res) => {
+  res.json({ message: "home page" });
+});
+
+//routes
+require("./routes/auth.routes")(app);
+require("./routes/user.routes")(app);
+
+// set port, listen for requests
+const PORT = process.env.PORT || 8080;
+app.listen(PORT, () => {
+  console.log(`Server is running on port ${PORT}.`);
+});
